pyproject.toml
[tool.poetry]
name = "imgvisint"
version = "0.1.0"
description = "Image Upscale/Interpolation visualisation"
authors = ["Zhdanov E.A. <your_email@example.com>"]  # Добавьте email в формате "Name <email>"
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"  # Версия Python указывается здесь, а не в requires-python
certifi = "==2025.1.31"
charset-normalizer = "==3.4.1"
colorama = "==0.4.6"
contourpy = "==1.3.1"
cycler = "==0.12.1"
filelock = "==3.18.0"
fonttools = "==4.56.0"
fsspec = "==2025.3.0"
h5py = "==3.13.0"
huggingface-hub = "==0.25.2"
idna = "==3.10"
jinja2 = "==3.1.6"
kiwisolver = "==1.4.8"
markupsafe = "==3.0.2"
matplotlib = "==3.10.1"
mpmath = "==1.3.0"
networkx = "==3.4.2"
numpy = "==1.25.2"
opencv-python = "==4.7.0.72"
packaging = "==24.2"
pillow = "==11.1.0"
pyparsing = "==3.2.1"
python-dateutil = "==2.9.0"
pyyaml = "==6.0.2"
requests = "==2.32.3"
scipy = "==1.9.3"
six = "==1.17.0"
super-image = "==0.1.7"
sympy = "==1.13.1"
torch = "==2.6.0"
torchvision = "==0.21.0"
tqdm = "==4.67.1"
typing-extensions = "==4.12.2"
urllib3 = "==2.3.0"
RealESRGAN = {git = "https://github.com/sberbank-ai/Real-ESRGAN.git", branch = "main"}
ruff = "^0.11.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Базовые настройки
line-length = 120
target-version = "py310"

lint.select = [
    "E",   # Ошибки (Pyflakes)
    "F",   # Контроль стиля (pycodestyle)
    "I",   # Сортировка импортов (isort)
    "B",   # Баги (flake8-bugbear)
    "C4",  # Сложность кода (mccabe)
    "UP",  # Модернизация кода (pyupgrade)
    "YTT", # Проверка типов в шаблонах
    "ANN", # Аннотации типов
    "S",   # Безопасность (bandit)
    "T20", # Медленные тесты (flake8-testing)
]
lint.ignore = [
    "D",       # Документация (pydocstyle) - часто отключают
    "PL",      # Pylint-правила (слишком строгие)
    "ERA001",  # Комментарии с лишними символами
    "F401",    # Названия импортов должны отличаться для перегруженных классов
]
lint.fixable = ["ALL"]

[lint.isort]
known-first-party = ["imgvisint"]  # Укажите имя вашего пакета

[lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"
docstring-quotes = "double"

[lint.per-file-ignores]
# Игнорирование конкретных файлов
"tests/*" = ["ANN", "S"]